requirements.txt
Flask==3.0.3
Bootstrap_Flask==2.4.0
WTForms==3.1.2
Flask_WTF==1.2.1
Werkzeug==3.0.4
flask_sqlalchemy==3.1.1
SQLAlchemy~=2.0.35
python-dotenv~=1.0.1